+++ app/src/main/java/com/falconssoft/woodysystem/LoadingOrder.java
@@ -392,11 +392,11 @@ public class LoadingOrder extends AppCompatActivity {
                                 && innerObject.getString("LOCATION").equals(loc)) {
 
                             BundleInfo bundleInfo = new BundleInfo();
-                            bundleInfo.setThickness(innerObject.getInt("THICKNESS"));
-                            bundleInfo.setWidth(innerObject.getInt("WIDTH"));
-                            bundleInfo.setLength(innerObject.getInt("LENGTH"));
+                            bundleInfo.setThickness(innerObject.getDouble("THICKNESS"));
+                            bundleInfo.setWidth(innerObject.getDouble("WIDTH"));
+                            bundleInfo.setLength(innerObject.getDouble("LENGTH"));
                             bundleInfo.setGrade(innerObject.getString("GRADE"));
-                            bundleInfo.setNoOfPieces(innerObject.getInt("PIECES"));
+                            bundleInfo.setNoOfPieces(innerObject.getDouble("PIECES"));
                             bundleInfo.setBundleNo(innerObject.getString("BUNDLE_NO"));
                             bundleInfo.setLocation(innerObject.getString("LOCATION"));
                             bundleInfo.setArea(innerObject.getString("AREA"));
