+++ app/src/main/java/com/falconssoft/woodysystem/reports/LoadingOrderReport.java
@@ -797,11 +797,11 @@ public class LoadingOrderReport extends AppCompatActivity {
                         JSONObject finalObject = parentArrayOrders.getJSONObject(i);
 
                         Orders order = new Orders();
-                        order.setThickness(finalObject.getInt("THICKNESS"));
-                        order.setWidth(finalObject.getInt("WIDTH"));
-                        order.setLength(finalObject.getInt("LENGTH"));
+                        order.setThickness(finalObject.getDouble("THICKNESS"));
+                        order.setWidth(finalObject.getDouble("WIDTH"));
+                        order.setLength(finalObject.getDouble("LENGTH"));
                         order.setGrade(finalObject.getString("GRADE"));
-                        order.setNoOfPieces(finalObject.getInt("PIECES"));
+                        order.setNoOfPieces(finalObject.getDouble("PIECES"));
                         order.setBundleNo(finalObject.getString("BUNDLE_NO"));
                         order.setLocation(finalObject.getString("LOCATION"));
                         order.setArea(finalObject.getString("AREA"));
@@ -1099,11 +1099,11 @@ public class LoadingOrderReport extends AppCompatActivity {
                         JSONObject finalObject = parentArrayOrders.getJSONObject(i);
 
                         Orders order = new Orders();
-                        order.setThickness(finalObject.getInt("THICKNESS"));
-                        order.setWidth(finalObject.getInt("WIDTH"));
-                        order.setLength(finalObject.getInt("LENGTH"));
+                        order.setThickness(finalObject.getDouble("THICKNESS"));
+                        order.setWidth(finalObject.getDouble("WIDTH"));
+                        order.setLength(finalObject.getDouble("LENGTH"));
                         order.setGrade(finalObject.getString("GRADE"));
-                        order.setNoOfPieces(finalObject.getInt("PIECES"));
+                        order.setNoOfPieces(finalObject.getDouble("PIECES"));
                         order.setBundleNo(finalObject.getString("BUNDLE_NO"));
                         order.setLocation(finalObject.getString("LOCATION"));
                         order.setArea(finalObject.getString("AREA"));
